Add in ground rules
We want to make sure that the first users of this website have somewher
e to start off. Of course with any kind of website like this. The users
will find their own way to police things.

For now I used an ordered list for this. Until we are ready for interac
tivity using Javascript.